Added a new test_conftest.py file with the temp directory fixture and added a test to check the ParaFrame type of the monkey-patched core file
