+++ tests/test_paraframe_e2e.py
@@ -1,22 +1,10 @@
-
 import os
 import shutil
 import pandas as pd
 import pytest
 import numpy as np
 from hallmark import ParaFrame
-
-
-@pytest.fixture
-def create_temp_data(tmp_path):
-    data_dir = tmp_path / "data"
-    for a in range(10):
-        subdir = data_dir / f"a_{a}"
-        subdir.mkdir(parents=True)
-        for b in range(10, 20):
-            (subdir / f"b_{b}.txt").touch()
-    return data_dir
-
+from test_conftest import create_temp_data
 
 def test_paraframe_class_functionality(create_temp_data):
     # a user wants to create a paraframe
